+++ src/templates/stacked_skeleton/main.ts
@@ -9,10 +9,8 @@ export class Main extends Base {
 
     this._dartString = `import 'package:flutter/material.dart';
 import 'package:logger/logger.dart';
+import 'package:${projectName}/src/app/app.dart';
 import 'package:${projectName}/src/app/generated/locator/locator.dart';
-import 'package:${projectName}/src/app/generated/router/router.gr.dart';
-import 'package:${projectName}/src/ui/global/app_colors.dart';
-import 'package:stacked_services/stacked_services.dart';
 
 main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -24,24 +22,7 @@ main() {
   setupLocator();
 
   // Runs the app :)
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: '${projectName}',
-      initialRoute: Routes.startupViewRoute,
-      onGenerateRoute: Router().onGenerateRoute,
-      navigatorKey: locator<NavigationService>().navigatorKey,
-      theme: ThemeData(
-        primaryColor: primaryColor,
-        accentColor: accentColor,
-      ),
-    );
-  }
+  runApp(App());
 }`;
   }