+++ src/templates/stacked_skeleton/ui/views/startup_view.ts
@@ -1,5 +1,5 @@
-import { Base } from "../../../base";
-import { YamlHelper } from "../../../../../utils/yaml_helper";
+import { Base } from "../../base";
+import { YamlHelper } from "../../../../utils/yaml_helper";
 
 export class StartupView extends Base {
   private _dartString: string;
@@ -9,8 +9,8 @@ export class StartupView extends Base {
 
     this._dartString = `import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
+import 'package:${YamlHelper.getProjectName()}/src/ui/widgets/dumb/skeleton.dart';
 import 'package:${YamlHelper.getProjectName()}/src/ui/global/ui_helpers.dart';
-import 'package:${YamlHelper.getProjectName()}/src/ui/global/app_colors.dart';
 import 'package:${YamlHelper.getProjectName()}/src/ui/views/startup/startup_view_model.dart';
 
 class StartupView extends StatefulWidget {
@@ -29,24 +29,17 @@ class _StartupViewState extends State<StartupView> {
         StartupViewModel model,
         Widget child,
       ) {
-        return Scaffold(
+        return Skeleton(
+          isBusy: model.isBusy,
           body: Center(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
               crossAxisAlignment: CrossAxisAlignment.center,
               children: <Widget>[
                 // TODO(yazeed): Put Your Logo Here :)
-                // Image.asset(
-                //   'assets/icons/logo.png',
-                //   width: blockSizeHorizontal(context) * 25,
-                // ),
-            
-                SizedBox(
-                  height: blockSizeHorizontal(context) * 10,
-                ),
-                CircularProgressIndicator(
-                  valueColor: AlwaysStoppedAnimation(primaryColor),
-                ),
+
+                verticalSpaceMedium(context),
+                CircularProgressIndicator(),
               ],
             ),
           ),