+++ src/utils/create_directory.ts
@@ -2,10 +2,11 @@ import * as mkdirp from "mkdirp";
 
 export function createDirectory(targetDirectory: string): Promise<void> {
   return new Promise((resolve, reject) => {
-    mkdirp(targetDirectory, (error) => {
-      if (error) {
-        return reject(error);
+    mkdirp(targetDirectory).then((made) => {
+      if (!made) {
+        return reject("Didn't make directory!");
       }
+
       resolve();
     });
   });