+++ src/utils/view_file.ts
@@ -2,12 +2,11 @@ import * as path from "path";
 import * as _ from "lodash";
 import { existsSync } from "fs";
 import { FileSystemManager } from "./file_system_manager";
-import { YamlHelper } from "./yaml_helper";
 import { Utils } from "./utils";
-import { View } from "../templates/stacked_skeleton/app/ui/views/view";
-import { ViewModel } from "../templates/stacked_skeleton/app/ui/views/view_model";
-import { StartupViewModel } from "../templates/stacked_skeleton/app/ui/views/startup_view_model";
-import { StartupView } from "../templates/stacked_skeleton/app/ui/views/startup_view";
+import { View } from "../templates/stacked_skeleton/ui/views/view";
+import { ViewModel } from "../templates/stacked_skeleton/ui/views/view_model";
+import { StartupViewModel } from "../templates/stacked_skeleton/ui/views/startup_view_model";
+import { StartupView } from "../templates/stacked_skeleton/ui/views/startup_view";
 
 export class ViewFile {
   constructor(